src/alphabrain/pyproject.toml
[tool.poetry]
name = "alphabrain"
version = "0.1.0"
description = ""
authors = ["Sruinard <41334908+Sruinard@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "~3.8"
fastapi = {extras = ["uvicorn"], version = "^0.85.0"}
uvicorn = "^0.18.3"
numpy = "^1.23.3"
ipykernel = "^6.15.3"
jax = "0.3.15"
jaxlib = "0.3.15"
flax = "0.5.0"
clu = "^0.0.7"
python-dotenv = "^0.21.0"
azure-storage-file-share = "^12.9.0"
azure-storage-queue = "^12.4.0"
azure-ai-ml = "^1.0.0"
azure-identity = "^1.11.0"
dm-sonnet = "^2.0.0"
mldesigner = "^0.1.0b8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
